requirements.txt
pyaudio
vosk
pynput
python-dotenv
